+++ packages/frontend/src/components/UserSettings/AppearanceSettingsPanel/CreatePaletteModal.tsx
@@ -1,3 +1,4 @@
+import { ECHARTS_DEFAULT_COLORS } from '@lightdash/common';
 import { type ModalProps } from '@mantine/core';
 import { type FC } from 'react';
 import {
@@ -6,9 +7,27 @@ import {
 } from '../../../hooks/appearance/useOrganizationAppearance';
 import useToaster from '../../../hooks/toaster/useToaster';
 import { PaletteModalBase, type PaletteFormValues } from './PaletteModalBase';
-import { PRESET_COLOR_PALETTES } from './palettes';
 
 type Props = Pick<ModalProps, 'opened' | 'onClose'>;
+
+const DEFAULT_COLOR_PALETTE = {
+    name: 'Default',
+    colors: [
+        // Use the initial 9 colors directly from ECHARTS to keep them in sync:
+        ...ECHARTS_DEFAULT_COLORS,
+        '#33ff7d',
+        '#33ffb1',
+        '#33ffe6',
+        '#33e6ff',
+        '#33b1ff',
+        '#337dff',
+        '#3349ff',
+        '#5e33ff',
+        '#9233ff',
+        '#c633ff',
+        '#ff33e1',
+    ],
+};
 
 export const CreatePaletteModal: FC<Props> = ({ opened, onClose }) => {
     const { data: palettes = [] } = useColorPalettes();
@@ -48,7 +67,7 @@ export const CreatePaletteModal: FC<Props> = ({ opened, onClose }) => {
             isLoading={createColorPalette.isLoading}
             initialValues={{
                 name: '',
-                colors: PRESET_COLOR_PALETTES[0].colors,
+                colors: DEFAULT_COLOR_PALETTE.colors,
             }}
             title="Create new palette"
             submitButtonText="Create palette"
