+++ packages/frontend/src/components/UserSettings/AppearanceSettingsPanel/EditPaletteModal.tsx
@@ -5,7 +5,6 @@ import {
     useColorPalettes,
     useUpdateColorPalette,
 } from '../../../hooks/appearance/useOrganizationAppearance';
-import useToaster from '../../../hooks/toaster/useToaster';
 import { PaletteModalBase, type PaletteFormValues } from './PaletteModalBase';
 
 type EditPaletteModalProps = Pick<ModalProps, 'opened' | 'onClose'> & {
@@ -19,32 +18,15 @@ export const EditPaletteModal: FC<EditPaletteModalProps> = ({
 }) => {
     const { data: palettes = [] } = useColorPalettes();
     const updateColorPalette = useUpdateColorPalette();
-    const { showToastSuccess, showToastApiError } = useToaster();
 
     const handleUpdatePalette = (values: PaletteFormValues) => {
         if (!values.name) return;
 
-        updateColorPalette.mutate(
-            {
-                uuid: palette.colorPaletteUuid,
-                name: values.name,
-                colors: values.colors,
-            },
-            {
-                onSuccess: (updatedPalette) => {
-                    onClose();
-                    showToastSuccess({
-                        title: `Palette "${updatedPalette.name}" updated successfully`,
-                    });
-                },
-                onError: (error) => {
-                    showToastApiError({
-                        title: 'Failed to update palette',
-                        apiError: error.error,
-                    });
-                },
-            },
-        );
+        updateColorPalette.mutate({
+            uuid: palette.colorPaletteUuid,
+            name: values.name,
+            colors: values.colors,
+        });
     };
 
     // Filter out the current palette from the existing names to avoid self-comparison
